requirements.txt
fastapi==0.111.0
PyGithub==2.3.0
Authlib==1.3.1
itsdangerous==2.2.0